10.16

+++ backjoon Q1~7/src/main/java/com/cwc/study/group3/day1/chap1.java
@@ -24,7 +24,8 @@ public class chap1 {
 
         /*5) A/B=? (0 < A, B < 10)*/
 
-        System.out.println(A/B);
+        System.out.println((double) A/(double) B);
+
 
         /*6) A+B,A-B,A*B,A/B,A%B (1 ≤ A, B ≤ 10,000)*/
         int plus = A+B;
